requirements.txt
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0

